requirements.txt
pylint
pylint-django
requests
django
numpy
pillow
piexif
h5py


